fixed ExceptionInitializerError on hibernate mappings

+++ src/main/java/gr/teicm/offersaroundws/controller/OfferstempController.java
@@ -13,7 +13,7 @@ public class OfferstempController {
       this.view = view;
    }
     
-    public Short getBusinessId(){
+    public int getBusinessId(){
        return model.getBusinessId();
    }
